distributed classes into packages in the service, extract calculateAverageRating to new  ExtensionService and added to context
