+++ src/main/java/servlet/RegistrationServlet.java
@@ -11,7 +11,7 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import listener.ContextListener;
-import service.UserService;
+import service.user.UserService;
 import service.registration.RegistrationDataExtractor;
 
 import java.io.IOException;
