+++ src/main/java/servlet/StatisticServlet.java
@@ -10,7 +10,7 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import listener.ContextListener;
-import service.UserService;
+import service.user.UserService;
 
 import java.io.IOException;
 import java.util.List;
